+++ src/hooks/useAnimatedPresence/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./useAnimatedPresence";
+export type { AnimatedPresenceResult } from "./types";
