+++ src/hooks/useAnimatedPresence/types.ts
@@ -1,0 +1,4 @@
+export type AnimatedPresenceResult = {
+  shouldRender: boolean;
+  isExiting: boolean;
+};
