+++ src/views/Play/components/Board/Row/index.ts
@@ -1,0 +1,1 @@
+export { Row } from "./Row";
