+++ src/views/Play/components/Board/Tile/index.ts
@@ -1,0 +1,1 @@
+export { Tile } from "./Tile";
