+++ src/views/Play/components/Board/useBoardController/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./useBoardController";
