+++ src/views/Play/components/Board/useBoardController/useBoardController.ts
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from "react";
 import { buildBoardRows } from "@domain/wordle";
-import { BOARD_SHAKE_DURATION_MS } from "./constants";
-import type { BoardRowViewModel, UseBoardControllerParams } from "./types";
+import { BOARD_SHAKE_DURATION_MS } from "../constants";
+import type { BoardRowViewModel, UseBoardControllerParams } from "../types";
 
 const useBoardController = ({
   guesses,
