+++ src/views/Play/components/Dialogs/DeveloperConsoleDialog/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DeveloperConsoleDialog";
