+++ src/views/Play/components/Dialogs/SettingsHint/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SettingsHint";
