+++ src/views/Play/sections/BoardSection/BoardContent/index.ts
@@ -1,0 +1,1 @@
+export { BoardContent } from "./BoardContent";
