+++ src/views/Play/sections/BoardSection/HardModeProgressBar/index.ts
@@ -1,0 +1,1 @@
+export { HardModeProgressBar } from "./HardModeProgressBar";
