+++ src/views/Play/sections/DialogsSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DialogsSection";
