+++ src/views/Play/sections/KeyboardSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./KeyboardSection";
