+++ src/views/Play/sections/PlayContent/PlayContent.tsx
@@ -1,10 +1,10 @@
 import type { JSX } from "react";
 import { useFeatureFlags } from "@providers/FeatureFlags";
-import BoardSection from "./BoardSection/BoardSection";
-import DialogsSection from "./DialogsSection";
-import KeyboardSection from "./KeyboardSection";
-import SettingsDrawer from "./SettingsDrawer";
-import Toolbar from "./Toolbar/Toolbar";
+import BoardSection from "../BoardSection/BoardSection";
+import DialogsSection from "../DialogsSection";
+import KeyboardSection from "../KeyboardSection";
+import SettingsDrawer from "../SettingsDrawer";
+import Toolbar from "../Toolbar/Toolbar";
 
 export const PlayContent = (): JSX.Element => {
   const { settingsDrawerEnabled } = useFeatureFlags();
