+++ src/views/Play/sections/PlayContent/index.ts
@@ -1,0 +1,1 @@
+export { PlayContent } from "./PlayContent";
