+++ src/views/Play/sections/SettingsDrawer/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SettingsDrawer";
