+++ src/views/Play/sections/Toolbar/HardModeTimerIndicator/index.ts
@@ -1,0 +1,1 @@
+export { HardModeTimerIndicator } from "./HardModeTimerIndicator";
